requirements/testing.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
astroid==2.4.2            # via pylint, pylint-celery
attrs==20.3.0             # via pytest
boto3==1.16.45            # via -r requirements/base.txt
botocore==1.19.45         # via -r requirements/base.txt, boto3, s3transfer
certifi==2020.12.5        # via -r requirements/base.txt, requests
chardet==4.0.0            # via -r requirements/base.txt, requests
click-log==0.3.2          # via edx-lint
click==7.1.2              # via -r requirements/base.txt, click-log, edx-lint
coverage==5.3.1           # via -r requirements/testing.in, pytest-cov
ddt==1.4.1                # via -r requirements/testing.in
deprecated==1.2.10        # via -r requirements/base.txt, pygithub
edx-lint==1.6             # via -r requirements/testing.in
gitdb==4.0.5              # via -r requirements/base.txt, gitpython
gitpython==3.1.11         # via -r requirements/base.txt
httpretty==1.0.3          # via -r requirements/testing.in
idna==2.10                # via -r requirements/base.txt, requests
iniconfig==1.1.1          # via pytest
isort==5.6.4              # via pylint
jmespath==0.10.0          # via -r requirements/base.txt, boto3, botocore
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via pylint
mock==4.0.3               # via -r requirements/testing.in
packaging==20.8           # via pytest
pluggy==0.13.1            # via pytest
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via -r requirements/testing.in
pygithub==1.54.1          # via -r requirements/base.txt
pyjwt==1.7.1              # via -r requirements/base.txt, pygithub
pylint-celery==0.3        # via edx-lint
pylint-django==2.3.0      # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.6.0             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements/testing.in
pytest==6.2.1             # via -r requirements/testing.in, pytest-cov
python-dateutil==2.8.1    # via -r requirements/base.txt, botocore
requests==2.25.1          # via -r requirements/base.txt, pygithub
s3transfer==0.3.3         # via -r requirements/base.txt, boto3
six==1.15.0               # via -r requirements/base.txt, astroid, edx-lint, python-dateutil
smmap==3.0.4              # via -r requirements/base.txt, gitdb
testfixtures==6.17.0      # via -r requirements/testing.in
toml==0.10.2              # via pylint, pytest
urllib3==1.26.2           # via -r requirements/base.txt, botocore, requests
wrapt==1.11.2             # via -c requirements/constraints.txt, -r requirements/base.txt, astroid, deprecated
